+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - added more debugging output
 - added `--analysis_keep_files` option for debugging purposes, which won't delete the output
   of the image analysis framework
+- added options for defining a frame window (`--from_frame`, `--to_frame`)
 
 
 0.0.1 (2021-07-21)